Great benchmarking tool, proper handling of any graph size, made p work

+++ src/johnson.h
@@ -32,7 +32,7 @@ typedef struct Graph {
 } Graph_t; */
 
 
-inline int* johnson_init(const int n, const int p, const unsigned long seed) {
+inline int* johnson_init(const int n, const double p, const unsigned long seed) {
   static std::uniform_real_distribution<double> flip(0, 1);
   // TODO: create negative edges without negative cycles
   static std::uniform_int_distribution<int> choose_weight(1, 100);
